requirements.txt
cloudinary==1.44.1
fastapi==0.119.1
firebase-admin==7.1.0
pymongo==4.15.3
pyrebase4==4.8.0
python-dotenv==1.1.1
python-multipart==0.0.20
uvicorn==0.38.0
